Current weather UI & Preferences

+++ app/src/main/java/com/example/forecast/data/db/entity/CurrentWeatherEntry.kt
@@ -9,14 +9,14 @@ const val CURRENT_WEATHER_ID = 0
 @Entity(tableName = "current_weather")
 data class CurrentWeatherEntry(
         @SerializedName("feelslike")
-        val feelsLike: Int,
+        val feelsLike: Double,
         @SerializedName("is_day")
         val isDay: String,
-        val precip: Int,
-        val temperature: Int,
+        val precip: Double,
+        val temperature: Double,
         @SerializedName("uv_index")
-        val uvIndex: Int,
-        val visibility: Int,
+        val uvIndex: Double,
+        val visibility: Double,
         @SerializedName("weather_code")
         val weatherCode: Int,
         @SerializedName("weather_descriptions")
@@ -26,7 +26,7 @@ data class CurrentWeatherEntry(
         @SerializedName("wind_dir")
         val windDir: String,
         @SerializedName("wind_speed")
-        val windSpeed: Int
+        val windSpeed: Double
 ){
     @PrimaryKey(autoGenerate = false)
     var id: Int = CURRENT_WEATHER_ID
